# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API Cửa hàng quần áo'
  description: 'Tài liệu API cho hệ thống cửa hàng quần áo'
  version: 1.0.0
paths:
...
  /api/categories:
    get:
      tags:
        - 'Danh mục'
      summary: 'Danh sách danh mục'
      description: 'Lấy danh sách danh mục kèm các danh mục con. Truyền per_page=0 để lấy toàn bộ.'
      operationId: listCategories
      parameters:
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
            minimum: 0
          example: 20
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
      responses:
        200:
          description: 'Lấy danh sách danh mục thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/CategoryDetail'
                      pagination:
                        properties:
                          page:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 20
                          totalItems:
                            type: integer
                            example: 15
                          totalPages:
                            type: integer
                            example: 1
                        type: object
                        nullable: true
                    type: object
                type: object
        422:
          description: 'Dữ liệu phân trang không hợp lệ'
    post:
      tags:
        - 'Danh mục'
      summary: 'Tạo danh mục'
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Áo thun'
                parent_id:
                  type: integer
                  example: null
                  nullable: true
                image:
                  description: 'URL ảnh danh mục. Nếu không gửi hoặc gửi null/rỗng, database sử dụng ảnh mặc định.'
                  type: string
                  format: uri
                  example: 'https://cdn.example.com/categories/ao-thun.jpg'
                  nullable: true
              type: object
      responses:
        201:
          description: 'Tạo danh mục thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 201
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    properties:
                      items:
                        $ref: '#/components/schemas/CategorySummary'
                      pagination:
                        type: object
                        example: null
                        nullable: true
                    type: object
                type: object
        401:
          description: 'Chưa xác thực'
        403:
          description: 'Chỉ admin được phép thao tác'
        422:
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearerAuth: []
  '/api/categories/{id}':
    get:
      tags:
        - 'Danh mục'
      summary: 'Chi tiết danh mục'
      operationId: getCategory
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: 'Lấy thông tin danh mục thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    properties:
                      items:
                        $ref: '#/components/schemas/CategoryDetail'
                      pagination:
                        type: object
                        example: null
                        nullable: true
                    type: object
                type: object
        404:
          description: 'Không tìm thấy danh mục'
    put:
      tags:
        - 'Danh mục'
      summary: 'Cập nhật danh mục'
      operationId: updateCategory
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Áo sơ mi'
                parent_id:
                  type: integer
                  example: null
                  nullable: true
                image:
                  description: 'URL ảnh mới. Nếu không gửi hoặc gửi null/rỗng, ảnh hiện tại được giữ nguyên.'
                  type: string
                  format: uri
                  example: 'https://cdn.example.com/categories/ao-so-mi.jpg'
                  nullable: true
              type: object
      responses:
        200:
          description: 'Cập nhật danh mục thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    properties:
                      items:
                        $ref: '#/components/schemas/CategoryDetail'
                      pagination:
                        type: object
                        example: null
                        nullable: true
                    type: object
                type: object
        401:
          description: 'Chưa xác thực'
        403:
          description: 'Chỉ admin được phép thao tác'
        404:
          description: 'Không tìm thấy danh mục'
        422:
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Danh mục'
      summary: 'Xóa mềm danh mục'
      description: 'Đánh dấu danh mục đã xóa và đưa các danh mục con trực tiếp của nó về cấp gốc.'
      operationId: deleteCategory
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: 'Xóa danh mục thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    type: object
                    example: null
                    nullable: true
                type: object
        401:
          description: 'Chưa xác thực'
        403:
          description: 'Chỉ admin được phép thao tác'
        404:
          description: 'Không tìm thấy danh mục'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    CategorySummary:
      required:
        - id
        - name
        - image
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Áo thun'
        image:
          type: string
          format: uri
          example: 'https://cdn.example.com/categories/ao-thun.jpg'
        parent_id:
          type: integer
          example: null
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    CategoryDetail:
      required:
        - id
        - name
        - image
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Áo thun'
        image:
          type: string
          format: uri
          example: 'https://cdn.example.com/categories/ao-thun.jpg'
        parent_id:
          type: integer
          example: null
          nullable: true
        parent:
          oneOf:
            - $ref: '#/components/schemas/CategorySummary'
          nullable: true
        children:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Nhập JWT token theo dạng Bearer token.'
      bearerFormat: JWT
      scheme: bearer
...